extract GrantStat into separate file
